+++ lib/haml_coffee_assets/action_view/template_handler.rb
@@ -7,6 +7,10 @@ module HamlCoffeeAssets
         new(template).render
       end
 
+      def self.stale?(last_compile_time)
+        GlobalContext.mtime > last_compile_time
+      end
+
       def initialize(template, partial = false, dependencies = [])
         @template     = template
         @partial      = partial
@@ -48,7 +52,7 @@ module HamlCoffeeAssets
       end
 
       def helpers
-        ::HamlCoffeeAssets.helpers
+        GlobalContext.to_s
       end
 
       def compiled_template