+++ lib/haml_coffee_assets/rails/engine.rb
@@ -33,6 +33,50 @@ module HamlCoffeeAssets
           end
         end
 
+        if ::Rails.env == "development"
+          # Monkey patch rails so it busts the server cache for templates
+          # depending on the global_context_asset.
+          #
+          # Currently, the only way to force rails to recompile a server template is to
+          # touch it. This is problematic because when the global_context_asset
+          # changes we need to manually touch every template that uses the congtext
+          # in some way.
+          #
+          # To ease development, make rails 'touch' and recompile hamlc templates
+          # when the global context has changed.
+          #
+          # Do this ONLY in development.
+          #
+          # TODO: Don't monkey patch rails.
+          module ::ActionView
+            class Template
+              def stale?
+                return false unless ::Rails.env == "development"
+                return false unless handler.respond_to?(:stale?)
+                handler.stale?(updated_at)
+              end
+
+              alias_method :old_render, :render
+
+              # by default, rails will only compile a template once
+              # path render so it recompiles the template if 'stale'
+              def render(view, locals, buffer=nil, &block)
+                if @compiled and stale?
+                  now = Time.now
+                  File.utime(now, now, identifier) # touch file
+                  ::Rails.logger.info "Busted cache for #{identifier} by touching it"
+
+                  view = refresh(view)
+                  @source = view.source
+                  @compiled = false
+                end
+                old_render(view, locals, buffer, &block)
+              end
+
+            end
+          end
+        end
+
         next unless app.assets
 
         # Register Tilt template (for Sprockets)